+++ realm/realm-library/src/main/java/io/realm/RealmResults.java
@@ -20,6 +20,7 @@ package io.realm;
 import android.os.Looper;
 
 import io.realm.internal.Collection;
+import io.realm.internal.SortDescriptor;
 import rx.Observable;
 
 /**
@@ -224,5 +225,34 @@ public class RealmResults<E extends RealmModel> extends OrderedRealmCollectionIm
         } else {
             throw new UnsupportedOperationException(realm.getClass() + " does not support RxJava.");
         }
+    }
+
+    /**
+     * @deprecated use {@link RealmQuery#distinct(String)} on the return value of {@link #where()} instead. This will
+     * be removed in coming 3.x.x minor releases.
+     */
+    @Deprecated
+    public RealmResults<E> distinct(String fieldName) {
+        SortDescriptor distinctDescriptor = SortDescriptor.getInstanceForDistinct(collection.getTable(), fieldName);
+        Collection distinctCollection = collection.distinct(distinctDescriptor);
+        return createLoadedResults(distinctCollection);
+    }
+
+    /**
+     * @deprecated use {@link RealmQuery#distinctAsync(String)} on the return value of {@link #where()} instead. This
+     * will be removed in coming 3.x.x minor releases.
+     */
+    @Deprecated
+    public RealmResults<E> distinctAsync(String fieldName) {
+        return where().distinctAsync(fieldName);
+    }
+
+    /**
+     * @deprecated use {@link RealmQuery#distinct(String, String...)} on the return value of {@link #where()} instead.
+     * This will be removed in coming 3.x.x minor releases.
+     */
+    @Deprecated
+    public RealmResults<E> distinct(String firstFieldName, String... remainingFieldNames) {
+        return where().distinct(firstFieldName, remainingFieldNames);
     }
 }