+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 * `RealmResults.removeChangeListeners()`. Use `RealmResults.removeAllChangeListeners()` instead.
 * `RealmObject.removeChangeListeners()`. Use `RealmObject.removeAllChangeListeners()` instead.
 
+### Deprecated
+
+* `RealmResults.distinct()` and `RealmResults.distinctAsync()`. Use `RealmQuery.distinct()` and `RealmQuery.distinctAsync()` instead.
+
 ### Enhancements
 
 * Added support for sorting by link's field (#672).